Organized dir structure and CMake files + refactoring
Made the SDL dependent parts into a separate lib module. It should be easy to create new contexts for other systems than SDL.
Context thus holds a buffer, which is the 'main' buffer. The specialization of the context then decides how it will be rendered.
The SDL context is quite basic and will redraw the full buffer to an SDL_Surface.
Also deleted the window class for now. Higher functionality will be put into its own library later on.